git-svn-id: svn://192.168.1.132/autoStock/autoStock@765 33c8eb2e-3fb0-4f7a-9790-3894f54e5fb4

+++ src/com/autoStock/signal/NormalizeInterface.java
@@ -5,5 +5,5 @@ package com.autoStock.signal;
  *
  */
 public interface NormalizeInterface {
-	public int normalize(double input);
+	public double normalize(double input);
 }
